move ColorScheme to constants
